requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.1
pymysql==1.1.1
bcrypt==4.1.3
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
python-dateutil==2.9.0
cryptography==42.0.8
gunicorn==21.2.0
waitress==3.0.0
